4-Bit LFSR with verification file
LFSR as needed is coded in the "design.sv", the required module is used and the basic logic i.e. finding XOR of bit-1 and bit-3 in a 4 bit code. Prevented the case of "0000" stuck loop by starting it with a bit set having at least a bit as "1". Can be tested by the testbench. Verified and tested by me using testbench simulation. Ready for review and integration!